requirements.txt
streamlit
pandas
fpdf
python-dotenv
openai
gensim
feedparser
sentiment_analysis_lib  # ton analyseur
